pagina/auth_api/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.10
Flask-CORS==4.0.0
gunicorn==20.1.0